+++ CHANGELOG.md
@@ -14,6 +14,14 @@ All notable changes to this RDK Service will be documented in this file.
 
 * Changes in CHANGELOG should be updated when commits are added to the main or release branches. There should be one CHANGELOG entry per JIRA Ticket. This is not enforced on sprint branches since there could be multiple changes for the same JIRA ticket during development. 
 
+## [1.12.0] - 2025-12-23
+### Fixed
+- Fixed the Ethernet interface enablement upon EntOS migration
+
+## [1.11.0] - 2025-12-19
+### Added
+- Added support to build for Thunder R5.x
+
 ## [1.10.0] - 2025-12-05
 ### Fixed
 - Fixed the GetIPSettings for RDK Backend to return only when all the information like DNS info retrieved